revamp to match requested api

+++ comsoc/approval.go
@@ -1,16 +1,16 @@
 package comsoc
 
-func ApprovalSWF(p Profile, thresholds []int) (count Count, err error) {
+func ApprovalSWF(p Profile, thresholds []int64) (count Count, err error) {
 	count = make(Count)
 	for i, v := range p {
-		for j := 0; j < thresholds[i]; j++ {
+		for j := int64(0); j < thresholds[i]; j++ {
 			count[v[j]]++
 		}
 	}
 	return count, nil
 }
 
-func ApprovalSCF(p Profile, thresholds []int) (bestAlts []Alternative, err error) {
+func ApprovalSCF(p Profile, thresholds []int64) (bestAlts []Alternative, err error) {
 	count, err := ApprovalSWF(p, thresholds)
 	return maxCount(count), err
 }
